Revert "fix compilation under ARM device. Apply patch #69659"
This reverts commit e0a2fcfb9f0bc75084e4bf01905524fd5e29edd6.
See further discussion on bug #296141
Basically float miss the needed precision.

+++ libs/widgets/graphicsview/regionframeitem.cpp
@@ -639,8 +639,8 @@ void RegionFrameItem::mouseMoveEvent(QGraphicsSceneMouseEvent* event)
 {
     const QSizeF maxSize = parentDImgItem()->boundingRect().size();
     const QPointF point  = mapToParent(event->pos());
-    qreal posX           = qBound(qreal(0.0), point.x(), maxSize.width());
-    qreal posY           = qBound(qreal(0.0), point.y(), maxSize.height());
+    qreal posX           = qBound(0., point.x(), maxSize.width());
+    qreal posY           = qBound(0., point.y(), maxSize.height());
     QRectF r             = rect();
 
     // Adjust edge
